Revert should specify encoding
At the moment the eWASM design only requires ABI encoding for the revert
call arguments. It is not enforced anywhere else in this specification.

We might want to support other types of arguments encoding other than
Solidity ABI encoding in the future, e.g. something based on varint. Future
proof revert so this is possible.

Additionally, it is much easier to provide the revert string as a plain
string rather than ABI encoding the "Error(string)" when generating
reverts in WASM.

Signed-off-by: Sean Young <sean@mess.org>